PIM-4937: Split LocaleListener and LocaleSubscriber
