+++ CHANGELOG-1.5.md
@@ -10,4 +10,4 @@
 - Change constructor of `Pim\Bundle\TranslationBundle\Twig\TranslationsExtension`.
   Replace `Oro\Bundle\LocaleBundle\Model\LocaleSettings` by `Symfony\Component\HttpFoundation\RequestStack`.
 - Change constructor of `Pim\Bundle\UserBundle\EventSubscriber\LocalSubscriber`.
-  Replace `Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage` by `Symfony\Component\HttpFoundation\RequestStack`
+  Rename by `Pim\Bundle\UserBundle\EventSubscriber\LocaleSubscriber` and remove `Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage`
